Link new wizard to the provider users index page
