+++ app/views/provider_interface/provider_users/index.html.erb
@@ -2,7 +2,11 @@
 
 <h1 class="govuk-heading-xl">Users</h1>
 
-<%= link_to 'Invite user', new_provider_interface_provider_user_path, class: 'govuk-button' %>
+<%= link_to(
+  'Invite user',
+  provider_interface_edit_invitation_basic_details_path,
+  class: 'govuk-button',
+) %>
 
 <% content_for :before_content do %>
   <div class="govuk-breadcrumbs">
