+++ spec/system/provider_interface/provider_invites_user_with_new_wizard_spec.rb
@@ -7,6 +7,9 @@ RSpec.feature 'Provider invites a new provider user using wizard interface' do
   scenario 'Provider sends invite to user' do
     FeatureFlag.activate(:providers_can_manage_users_and_permissions)
 
+    # TODO: test what happens with the feature flag off
+    # TODO: test what happens without necessary permissions
+
     given_i_am_a_provider_user_with_dfe_sign_in
     and_i_can_manage_applications_for_two_providers
     and_i_can_manage_users_for_a_provider
